requirements.txt
Flask==2.0.3
google-cloud-speech==2.9.0
google-cloud-texttospeech==2.5.0
